feat: 주류별 테이스팅노트 그룹화 로직을 위한 리팩토링
