+++ src/app/mypage/page.tsx
@@ -7,26 +7,15 @@ import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import LogoutIcon from "@mui/icons-material/Logout";
 import axios from "axios";
+import MyPageContentsComponent from "@/components/MyPageContentsComponent/MyPageContentsComponent";
 
 export default function MyPage() {
   const [isLoggedIn, setIsLoggedIn] = useState<Boolean | null>(null);
   const [currentUrl, setCurrentUrl] = useState<string>("");
   const [user, setUser] = useState<User | null>(null);
   const router = useRouter();
-  const handleLogout = async () => {
-    try {
-      await axios.post(
-        `${process.env.NEXT_PUBLIC_API_BASE_URL}/logout`,
-        {},
-        { withCredentials: true },
-      );
-      window.location.href = `${process.env.NEXT_PUBLIC_BASE_URL}`; // 로그아웃 후 메인 페이지로 이동
-    } catch (error) {
-      console.error("Error logging out", error);
-    }
-  };
+
   useEffect(() => {
     const checkAuth = async () => {
       try {
@@ -35,7 +24,7 @@ export default function MyPage() {
           {
             method: "GET",
             credentials: "include", // 쿠키 포함
-          },
+          }
         );
 
         if (response.status === 401) {
@@ -83,81 +72,6 @@ export default function MyPage() {
 
   // 페이지: 로그인 되어 있는 경우
   if (user) {
-    return (
-      <Stack sx={{ paddingTop: "10px" }}>
-        {/* 사용자 프로필 */}
-        <Box
-          sx={{
-            padding: "10px 15px",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            gap: "25px",
-          }}
-        >
-          <Image
-            src={user?.profileThumbnailImage || "default"}
-            alt="profile image"
-            width={70}
-            height={70}
-            style={{ borderRadius: "15px" }}
-          />
-          <Stack sx={{ gap: "5px" }}>
-            <Typography sx={{ fontSize: "18px" }}>
-              {user?.profileNickname}
-            </Typography>
-            {/*<Typography sx={{ fontSize: "15px", color: "gray" }}>*/}
-            {/*  지역*/}
-            {/*</Typography>*/}
-          </Stack>
-        </Box>
-
-        {/*<Button*/}
-        {/*  variant="contained"*/}
-        {/*  color="inherit"*/}
-        {/*  size="small"*/}
-        {/*  startIcon={<Edit fontSize="small" />}*/}
-        {/*  sx={{*/}
-        {/*    margin: "5px 15px",*/}
-        {/*    fontSize: "13px",*/}
-        {/*    color: "gray",*/}
-        {/*    backgroundColor: "#f5f5f5",*/}
-        {/*  }}*/}
-        {/*>*/}
-        {/*  회원 정보 수정*/}
-        {/*</Button>*/}
-        <Button
-          variant="contained"
-          color="inherit"
-          size="small"
-          startIcon={<LogoutIcon fontSize="small" />}
-          onClick={handleLogout} // 클릭 시 handleLogout 실행
-          sx={{
-            margin: "5px 15px",
-            fontSize: "13px",
-            color: "gray",
-            backgroundColor: "#f5f5f5",
-          }}
-        >
-          로그아웃
-        </Button>
-
-        <Divider sx={{ padding: "5px 0" }} />
-
-        <Typography
-          sx={{
-            paddingTop: "50px",
-            fontSize: "15px",
-            color: "gray",
-            textAlign: "center",
-          }}
-        >
-          내가 작성한 테이스팅 노트
-        </Typography>
-
-        {/* 사용자 활동 정보 */}
-        <UserTastingComponent userId={user?.userUuid} />
-      </Stack>
-    );
+    return <MyPageContentsComponent user={user} />;
   }
 }
